DDBTEAM-640: Moved vendor image processor into ampq
