+++ src/Message/VendorImageMessage.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * @file
+ */
+
+namespace App\Message;
+
+/**
+ * Class VendorImageMessage.
+ */
+class VendorImageMessage extends BaseMessage
+{
+}
