+++ src/MessageHandler/VendorImageMessageHandler.php
@@ -8,6 +8,8 @@ namespace App\Queue;
 
 use App\Entity\Source;
 use App\Entity\Vendor;
+use App\Message\SearchMessage;
+use App\Message\VendorImageMessage;
 use App\Service\VendorService\VendorImageValidatorService;
 use App\Utils\CoverVendor\VendorImageItem;
 use App\Utils\Message\ProcessMessage;
@@ -22,11 +24,13 @@ use Interop\Queue\Message;
 use Interop\Queue\Processor;
 use Karriere\JsonDecoder\JsonDecoder;
 use Psr\Log\LoggerInterface;
+use Symfony\Component\Messenger\Exception\UnrecoverableMessageHandlingException;
+use Symfony\Component\Messenger\Handler\MessageHandlerInterface;
 
 /**
- * Class CoverStoreProcessor.
+ * Class VendorImageMessageHandler.
  */
-class VendorImageProcessor implements Processor, TopicSubscriberInterface
+class VendorImageMessageHandler implements MessageHandlerInterface
 {
     private $em;
     private $imageValidator;
@@ -50,132 +54,95 @@ class VendorImageProcessor implements Processor, TopicSubscriberInterface
         $this->statsLogger = $statsLogger;
     }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function process(Message $message, Context $session)
+    public function __invoke(VendorImageMessage $message)
     {
-        $jsonDecoder = new JsonDecoder(true);
-        $processMessage = $jsonDecoder->decode($message->getBody(), ProcessMessage::class);
-
         // Look up vendor to get information about image server.
         $vendorRepos = $this->em->getRepository(Vendor::class);
-        $vendor = $vendorRepos->find($processMessage->getVendorId());
+        $vendor = $vendorRepos->find($message->getVendorId());
 
         // Look up source to get source url and link it to the image.
         $sourceRepos = $this->em->getRepository(Source::class);
         /** @var Source $source */
         $source = $sourceRepos->findOneBy([
-            'matchId' => $processMessage->getIdentifier(),
+            'matchId' => $message->getIdentifier(),
             'vendor' => $vendor,
         ]);
 
-        if ($source) {
-            switch ($processMessage->getOperation()) {
+        if (!is_null($source)) {
+            switch ($message->getOperation()) {
                 case VendorState::INSERT:
-                  $processorStatus = $this->processInsert($processMessage, $source);
-                  break;
+                    $this->processInsert($message, $source);
+                    break;
 
                 case VendorState::UPDATE:
-                  $processorStatus = $this->processUpdate($processMessage, $source);
-                  break;
+                    $this->processUpdate($message, $source);
+                    break;
 
                 default:
-                  $processorStatus = self::REJECT;
-          }
+                    throw new UnrecoverableMessageHandlingException('Unknown vendor operation');
+            }
         } else {
-            $processorStatus = self::REJECT;
+            throw new UnrecoverableMessageHandlingException('Source not found');
         }
-
-        return $processorStatus;
     }
 
     /**
      * Handle image inserts. Send update to cover store processor only if vendor image exists.
      *
-     * @param ProcessMessage $processMessage
+     * @param VendorImageMessage $message
      * @param Source $source
-     *
-     * @return string
      *
      * @throws GuzzleException
      */
-    private function processInsert(ProcessMessage $processMessage, Source $source): string
+    private function processInsert(VendorImageMessage $message, Source $source)
     {
         $item = new VendorImageItem();
         $item->setOriginalFile($source->getOriginalFile());
 
         $this->imageValidator->validateRemoteImage($item);
 
-        $processorStatus = self::REJECT;
-
         if ($item->isFound()) {
-            $this->producer->sendEvent('CoverStoreTopic', JSON::encode($processMessage));
+            $this->producer->sendEvent('CoverStoreTopic', JSON::encode($message));
 
             $source->setOriginalLastModified($item->getOriginalLastModified());
             $source->setOriginalContentLength($item->getOriginalContentLength());
-
-            $processorStatus = self::ACK;
+            $this->em->flush();
         } else {
             $source->setOriginalFile(null);
             $source->setOriginalLastModified(null);
             $source->setOriginalContentLength(null);
+            $this->em->flush();
 
             // Log that the image did not exists.
             $this->statsLogger->error('Vendor image error - not found', [
                 'service' => 'VendorImageProcessor',
-                'identifier' => $processMessage->getIdentifier(),
+                'identifier' => $message->getIdentifier(),
                 'url' => $source->getOriginalFile(),
             ]);
+
+            throw new UnrecoverableMessageHandlingException('Vendor image error - not found');
         }
-
-        $this->em->flush();
-
-        return $processorStatus;
     }
 
     /**
      * Handle image updates. Send update to cover store processor only if vendor image is updated.
      *
-     * @param ProcessMessage $processMessage
+     * @param VendorImageMessage $message
      * @param Source $source
-     *
-     * @return string
      *
      * @throws GuzzleException
      */
-    private function processUpdate(ProcessMessage $processMessage, Source $source): string
+    private function processUpdate(VendorImageMessage $message, Source $source)
     {
         $item = new VendorImageItem();
         $item->setOriginalFile($source->getOriginalFile());
 
         $this->imageValidator->isRemoteImageUpdated($item, $source);
 
-        $processorStatus = self::REJECT;
-
         if ($item->isUpdated()) {
-            $this->producer->sendEvent('CoverStoreTopic', JSON::encode($processMessage));
-
-            $processorStatus = self::ACK;
+            $this->producer->sendEvent('CoverStoreTopic', JSON::encode($message));
+        } else {
+            throw new UnrecoverableMessageHandlingException('Remote image is not updated');
         }
-
-        return $processorStatus;
     }
-
-    // phpcs:disable Symfony.Functions.ScopeOrder.Invalid
-
-    /**
-     * {@inheritdoc}
-     */
-    public static function getSubscribedTopics(): array
-    {
-        return [
-            'VendorImageTopic' => [
-                'processorName' => 'VendorImageProcessor',
-                'queueName' => 'CoverStoreQueue',
-            ],
-        ];
-    }
-
-    // phpcs:enable
 }
